Fix send_photo_file: reply-to targeting and send hanging (#251)
* Fix send_photo_file: reply to original message and prevent hanging

1. Add MessageId to ToolContext so tools know the original user message
2. Set MessageId in all LLM services (OpenAI, Ollama, Gemini) when creating ToolContext
3. SendPhotoToolService now defaults reply-to to the original user message
4. Read file into memory buffer before queuing to rate limiter (prevents file stream issues)
5. Add 120s timeout via CancellationToken to prevent indefinite hanging


* Add file size check (10MB limit) before loading into memory


* fix: change MessageId from int to long for Telegram API compatibility

---------

+++ TelegramSearchBot.Common/Model/AI/LlmContinuationSnapshot.cs
@@ -35,9 +35,9 @@ namespace TelegramSearchBot.Model.AI {
         public long ChatId { get; set; }
 
         /// <summary>
-        /// The original user message ID (for reply context, int for Telegram API compatibility)
+        /// The original user message ID (for reply context)
         /// </summary>
-        public int OriginalMessageId { get; set; }
+        public long OriginalMessageId { get; set; }
 
         /// <summary>
         /// The user ID who initiated the conversation
